+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Check if downloaded server is on latest version otherwise re-download it.
+
 ## 2.4.0
 
 - Support window/showDocument request, supporiting moving cursor after applying paredit command. #45